+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.4.0 / 2015-07-31
 ------------------
 - made `<pre>` text selectable / copyable
+- scrolling hack (view all content)
 
 0.3.0 / 2015-07-29
 ------------------